version 1.3 - added detection for clipping voltages added peak to peak calculator

+++ waveform.h
@@ -1,13 +1,13 @@
 
 typedef struct{
-    float timestamp;
-    float phaseA;
-    float phaseB;
-    float phaseC;
-    float line_current;
-    float frequency;
-    float power_factor;
-    float thd_percent;
+    double timestamp;
+    double phaseA;
+    double phaseB;
+    double phaseC;
+    double line_current;
+    double frequency;
+    double power_factor;
+    double thd_percent;
 } pql_data;
 
 double my_rms( int n, double sum);
